test: jpa 테스트 추가 및 fix : table sql 변경

+++ Pt-Batch/db/initdb.d/create_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `package`
 (
-    `package_seq`  int         NOT NULL AUTO_INCREMENT COMMENT '패키지 순번',
+    `package_seq`  bigint         NOT NULL AUTO_INCREMENT COMMENT '패키지 순번',
     `package_name` varchar(50) NOT NULL COMMENT '패키지 이름',
     `count`        int                  DEFAULT NULL COMMENT '이용권 수, NULL인 경우 무제한',
     `period`       int                  DEFAULT NULL COMMENT '기간(일), NULL인 경우 무제한',
@@ -11,8 +11,8 @@ CREATE TABLE `package`
 
 CREATE TABLE `pass`
 (
-    `pass_seq`        int         NOT NULL AUTO_INCREMENT COMMENT '이용권 순번',
-    `package_seq`     int         NOT NULL COMMENT '패키지 순번',
+    `pass_seq`        bigint         NOT NULL AUTO_INCREMENT COMMENT '이용권 순번',
+    `package_seq`     bigint         NOT NULL COMMENT '패키지 순번',
     `user_id`         varchar(20) NOT NULL COMMENT '사용자 ID',
     `status`          varchar(10) NOT NULL COMMENT '상태',
     `remaining_count` int                  DEFAULT NULL COMMENT '잔여 이용권 수, NULL인 경우 무제한',
@@ -26,8 +26,8 @@ CREATE TABLE `pass`
 
 CREATE TABLE `booking`
 (
-    `booking_seq`  int         NOT NULL AUTO_INCREMENT COMMENT '예약 순번',
-    `pass_seq`     int         NOT NULL COMMENT '이용권 순번',
+    `booking_seq`  bigint         NOT NULL AUTO_INCREMENT COMMENT '예약 순번',
+    `pass_seq`     bigint        NOT NULL COMMENT '이용권 순번',
     `user_id`      varchar(20) NOT NULL COMMENT '사용자 ID',
     `status`       varchar(10) NOT NULL COMMENT '상태',
     `used_pass`    tinyint(1) NOT NULL DEFAULT '0' COMMENT '이용권 사용 여부',
